Refactory changes (especially in Ming2) Changed timer to return time_t Added more catch cases Different interface to BEB Fixed warnings on windows Changed BEB constructor Partial porting to XMT Fixed bug for testing
git-svn-id: https://svn.vital-it.ch/svn/hp2c/trunk/Codeml_Variants/Fastcodeml@4970 95c5a404-1f88-0410-a6b0-c3f062b6f34a

+++ TreeAndSetsDependencies.cpp
@@ -147,7 +147,7 @@ unsigned int TreeAndSetsDependencies::measureRelativeEffort(void)
 //#endif
 		}
 	}
-	double time_leaf = timer.stop();
+	double time_leaf = static_cast<double>(timer.stop());
 
 	// Measure doTransition()
 	timer.start();
@@ -160,7 +160,7 @@ unsigned int TreeAndSetsDependencies::measureRelativeEffort(void)
 //			elementWiseMult(dummy, cf);
 //#endif
 		}
-	double time_non_leaf = timer.stop();
+	double time_non_leaf = static_cast<double>(timer.stop());
 	unsigned int effort_ratio = static_cast<unsigned int>(time_non_leaf/time_leaf+0.5);
 #else
 	unsigned int effort_ratio = 16;
